requirements.txt
beautifulsoup4==4.4.1
Django==1.8.5
django-allauth==0.23.0
lxml==3.4.4
requests==2.8.1